Move modules in `airflow.contrib.utils.log` to `airflow.utils.log`
