+++ airflow/contrib/utils/log/__init__.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -15,3 +14,13 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""
+This module is deprecated. Please use `airflow.utils.log`.
+"""
+
+import warnings
+
+warnings.warn(
+    "This module is deprecated. Please use `airflow.utils.log`.",
+    DeprecationWarning, stacklevel=2
+)
